[otbn] add more documentation in bignum lib
This documents the precomputation of the constant m0'
for Montgomery modular arithmetic.

Word-wise montgomery modular arithmetic requires
a quantity m0' to be precomputed once per modulus m.

This computation is based on the Dusse-Kaliski method
for computing the multiplicative modular inverse when
the modulus is of the form 2^k.
See DOI https://doi.org/10.1007/3-540-46877-3_21
section 3.2 (Algorithm "Modular Inverse" on p. 235)

Signed-off-by: Felix Miller <felix.miller@gi-de.com>